+++ app/Models/Book.php
@@ -6,5 +6,13 @@ use Illuminate\Database\Eloquent\Model;
 
 class Book extends Model
 {
-    //
+    //Relationships.
+    public function testament () {
+        return $this->belongsTo(Testament::class);
+    }
+
+    public function chapters () {
+        return $this->hasMany(Chapter::class);
+    }
+
 }