+++ app/Models/Testament.php
@@ -6,5 +6,8 @@ use Illuminate\Database\Eloquent\Model;
 
 class Testament extends Model
 {
-    //
+    //Relationships.
+    public function books () {
+        return $this->hasMany(Book::class);
+    }
 }